+++ forui/lib/src/widgets/alert.dart
@@ -139,7 +139,7 @@ extension type FAlertStyles(FVariants<FAlertVariantConstraint, FAlertVariant, FA
   }) {
     final primary = FAlertStyle(
       iconStyle: IconThemeData(color: colors.foreground, size: typography.md.fontSize),
-      titleTextStyle: typography.sm.copyWith(fontWeight: .w500, color: colors.foreground, leadingDistribution: .even),
+      titleTextStyle: typography.sm.copyWith(fontWeight: .w500, color: colors.foreground),
       subtitleTextStyle: (touch ? typography.xs : typography.sm).copyWith(color: colors.mutedForeground),
       decoration: ShapeDecoration(
         shape: RoundedSuperellipseBorder(